* Debugger: deleting a breakpoint leaves selection in the Breakpoints list * Debugger: changed "Load .DEB" checkbox caption to ".DEB files" to emphasize that it also affects saving .deb files, not just loading them * updated docs * Taseditor: fixed extremely rare LagLog bug in History.jump() and RegisterBranching * Taseditor: fixed RegisterLuaChange to also consider LagLog changes
[[Split portion of a mixed commit.]]

+++ src/drivers/win/taseditor/laglog.cpp
@@ -132,7 +132,7 @@ void LAGLOG::EraseFrame(int frame)
 }
 void LAGLOG::InsertFrame(int frame, bool lagFlag, int frames)
 {
-	if (frame < lag_log.size())
+	if (frame < (int)lag_log.size())
 		// insert
 		lag_log.insert(lag_log.begin() + frame, frames, (lagFlag) ? 1 : 0);
 	else
